ServerFinal/requirements.txt
aniso8601==3.0.2
blis==0.2.4
certifi==2019.3.9
chardet==3.0.4
Click==7.0
cymem==2.0.2
Flask==1.0.2
flask-cas-ng==1.1.0
flask-heroku==0.1.9
Flask-SocketIO==3.3.2
Flask-SQLAlchemy==2.3.2
graphene==2.1.3
graphql-core==2.1
graphql-relay==0.4.5
gunicorn==19.9.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
jsonschema==2.6.0
MarkupSafe==1.1.1
murmurhash==1.0.2
numpy==1.16.3
plac==0.9.6
preshed==2.0.1
promise==2.2.1
psycopg2==2.7.7
python-engineio==3.5.0
python-socketio==3.1.2
requests==2.21.0
Rx==1.6.1
six==1.12.0
spacy==2.1.3
SQLAlchemy==1.3.3
srsly==0.0.5
thinc==7.0.4
tqdm==4.31.1
urllib3==1.24.2
wasabi==0.2.1
Werkzeug==0.15.0
xmltodict==0.12.0
